[kup-autocomplete / showcase] Updated how the showcase has been divided concerning the autocomplete.
