+++ packages/ketchup/src/components/kup-autocomplete/readme.md
@@ -16,7 +16,7 @@
 | `minimumChars`      | `minimum-chars`      | The minimum number of chars to trigger the autocomplete                                                | `number`                                                                                                              | `3`                                                        |
 | `multipleSelection` | `multiple-selection` | Allows more than one option to be selected at the same time.                                           | `boolean`                                                                                                             | `false`                                                    |
 | `noItemsLabel`      | `no-items-label`     | Label shown when there are no items found with a given filter                                          | `string`                                                                                                              | `'There are no items which\nmatch with the given filter.'` |
-| `placeholder`       | `placeholder`        | The placeholder string to set to the input for the autocomplete                                        | `string`                                                                                                              | `'Scegli'`                                                 |
+| `placeholder`       | `placeholder`        | The placeholder string to set to the input for the autocomplete                                        | `string`                                                                                                              | `'Choose'`                                                 |
 | `showClearIcon`     | `show-clear-icon`    | Shows the icon to clear the input                                                                      | `boolean`                                                                                                             | `false`                                                    |
 | `showDropdownIcon`  | `show-dropdown-icon` | Shows icon to force the dropdown menu to be opened                                                     | `boolean`                                                                                                             | `false`                                                    |
 | `sortBy`            | `sort-by`            | If different than 'none', sorts displayed results accordingly to the order provided by sortOrder prop. | `AutocompleteSortBy.CODE \| AutocompleteSortBy.DESCRIPTION \| AutocompleteSortBy.NONE`                                | `AutocompleteSortBy.NONE`                                  |
